+++ src/br/br-masks.js
@@ -1,9 +1,11 @@
 var m = angular.module('ui.utils.masks.br', [
 	require('../helpers'),
-	require('./cpf-cnpj/cpf-cnpj')
 ])
 .directive('uiBrBoletoBancarioMask', require('./boleto-bancario/boleto-bancario'))
 .directive('uiBrCepMask', require('./cep/cep'))
+.directive('uiBrCnpjMask', require('./cnpj/cnpj'))
+.directive('uiBrCpfMask', require('./cpf/cpf'))
+.directive('uiBrCpfcnpjMask', require('./cpf-cnpj/cpf-cnpj'))
 .directive('uiBrIeMask', require('./inscricao-estadual/ie'))
 .directive('uiNfeAccessKeyMask', require('./nfe/nfe'))
 .directive('uiBrPhoneNumber', require('./phone/br-phone'));