+++ src/br/cpf/cpf.js
@@ -1,0 +1,19 @@
+var StringMask = require('string-mask');
+var BrV = require('br-validations');
+var maskFactory = require('mask-factory');
+
+var cpfPattern = new StringMask('000.000.000-00');
+
+module.exports = maskFactory({
+	clearValue: function(rawValue) {
+		return rawValue.replace(/[^\d]/g, '').slice(0, 11);
+	},
+	format: function(cleanValue) {
+		return (cpfPattern.apply(cleanValue) || '').trim().replace(/[^0-9]$/, '');
+	},
+	validations: {
+		cpf: function(value) {
+			return BrV.cpf.validate(value);
+		}
+	}
+});