+++ src/global/scientific-notation/scientific-notation.test.js
@@ -16,8 +16,8 @@ describe('ui-scientific-notation-mask', function() {
 		var maskedInput = TestUtil.compile('<input ng-model="maskedModel" ui-scientific-notation-mask>');
 		var maskedModel = maskedInput.controller('ngModel');
 
-		expect(maskedModel.$parsers.length).toBe(model.$parsers.length + 2);
-		expect(maskedModel.$formatters.length).toBe(model.$formatters.length + 2);
+		expect(maskedModel.$parsers.length).toBe(model.$parsers.length + 1);
+		expect(maskedModel.$formatters.length).toBe(model.$formatters.length + 1);
 	});
 
 	it('should format initial model values', function() {